+++ hpath-directory/hpath-directory.cabal
@@ -28,7 +28,6 @@ extra-tmp-files:
 
 extra-source-files:
   ./src/HsDirectoryConfig.h.in
-  ./src/System/Directory/AFP.hs
   ./src/System/Win32/WindowsFilePath/*.h
   CHANGELOG.md
 
@@ -56,7 +55,8 @@ library
       , unix-bytestring         >=0.3
 
   exposed-modules:
-    System.Directory.AFP
+    System.Directory.AbstractFilePath
+    System.Directory.HPath
     System.Directory.Errors
     System.Directory.Types
 
@@ -67,10 +67,12 @@ library
     , bytestring               >=0.10
     , exceptions               >=0.10
     , filepath                 >=2.0.0.0
+    , file-io
     , IfElse
+    , hpath
     , safe-exceptions          >=0.1
     , split
-    , streamly                 >=0.7
+    , streamly                 >=0.8.2
     , streamly-bytestring      >=0.1.2
     , time                     >=1.8
     , transformers
