--- hpath-io/TODO.md
@@ -1,6 +1,0 @@
-# TODO
-
-## Tests
-
-* `doesExist` not tested
-* `readFileStream` only implicitly tested by `readFile`
