--- hpath-io/hpath-io.cabal
@@ -1,47 +1,0 @@
-name:                hpath-io
-version:             0.14.2
-synopsis:            High-level IO operations on files/directories
-description:         High-level IO operations on files/directories, utilizing type-safe Paths
--- bug-reports:
-license:             BSD3
-license-file:        LICENSE
-author:              Julian Ospald <hasufell@posteo.de>
-maintainer:          Julian Ospald <hasufell@posteo.de>
-copyright:           Julian Ospald 2016
-category:            Filesystem
-build-type:          Simple
-cabal-version:       1.14
-tested-with:         GHC==7.10.3
-                   , GHC==8.0.2
-                   , GHC==8.2.2
-                   , GHC==8.4.4
-                   , GHC==8.6.5
-                   , GHC==8.8.1
-extra-source-files:  README.md
-                     CHANGELOG.md
-
-library
-  if os(windows)
-    build-depends: unbuildable<0
-    buildable: False
-  exposed-modules:   HPath.IO
-  build-depends:     base >= 4.8 && <5
-                   , filepath >= 2.0.0.0
-                   , bytestring >= 0.10.0.0
-                   , exceptions
-                   , hpath >= 0.13 && < 0.14
-                   , hpath-directory >= 0.15.2 && < 0.16
-                   , hpath-posix     >= 0.14.3 && < 0.15
-                   , safe-exceptions >= 0.1
-                   , streamly >= 0.7
-                   , time >= 1.8
-                   , unix >= 2.8
-  if !impl(ghc>=7.11)
-    build-depends:     transformers
-  hs-source-dirs:      src
-  default-language:    Haskell2010
-
-
-source-repository head
-  type:     git
-  location: https://github.com/hasufell/hpath
